refactor: rename package from md2tag to ecifl2html and update related documentation
