+++ src/tests/test_ ecifl2html.py
@@ -1,8 +1,8 @@
-""" Unit tests for the MarkdownToHTML class in md2tag.py """
+""" Unit tests for the MarkdownToHTML class in ecifl2html.py """
 
 import unittest
 
-from md2tag import md2tag as hw
+from ecifl2html import ecifl2html as hw
 
 
 class TestMarkdownToHTML(unittest.TestCase):
